Separate pdfbox stuff from application domain
