+++ src/main/java/com/vnaskos/pdfguru/execute/DocumentManager.java
@@ -1,0 +1,17 @@
+package com.vnaskos.pdfguru.execute;
+
+import com.vnaskos.pdfguru.input.items.InputItem;
+
+import java.io.IOException;
+
+public interface DocumentManager extends Finishable {
+
+    void openNewDocument();
+
+    void saveDocument(String path) throws IOException;
+
+    void closeDocument() throws IOException;
+
+    void addInputItem(InputItem inputItem, float compression) throws IOException;
+
+}
