+++ src/main/java/com/vnaskos/pdfguru/execute/Finishable.java
@@ -1,0 +1,7 @@
+package com.vnaskos.pdfguru.execute;
+
+public interface Finishable {
+
+    void notifyFinish();
+
+}
